requirements.txt
fastapi
uvicorn
pandas
python-dotenv
requests
google-generativeai
langchain
chromadb
sentence-transformers
pypdf
